app/requirements.txt
streamlit>=1.30
pandas
yfinance
pyyaml
requests
pytz
stocknews
ta
plotly
streamlit-aggrid
transformers
streamlit-autorefresh
plyer